refactor: move language selection logic into composable
